requirements.txt
coinbase==2.1.0
pandas==2.0.3
numpy==1.24.3
tensorflow==2.13.0
websocket-client==1.6.1
python-dotenv==1.0.0
aiohttp==3.8.5
requests==2.31.0
PyJWT==2.8.0
cryptography==42.0.2
python-telegram-bot==20.7
scikit-learn==1.4.0 